Dissolve `Jekyll::SiteData` module
