+++ lib/jekyll.rb
@@ -45,6 +45,8 @@ module Jekyll
   autoload :Collection,          "jekyll/collection"
   autoload :Configuration,       "jekyll/configuration"
   autoload :Convertible,         "jekyll/convertible"
+  autoload :DataDirectory,       "jekyll/data_directory"
+  autoload :DataFile,            "jekyll/data_file"
   autoload :Deprecator,          "jekyll/deprecator"
   autoload :Document,            "jekyll/document"
   autoload :EntryFilter,         "jekyll/entry_filter"
@@ -77,7 +79,6 @@ module Jekyll
   autoload :Renderer,            "jekyll/renderer"
   autoload :LiquidRenderer,      "jekyll/liquid_renderer"
   autoload :Site,                "jekyll/site"
-  autoload :SiteData,            "jekyll/site_data"
   autoload :StaticFile,          "jekyll/static_file"
   autoload :Stevenson,           "jekyll/stevenson"
   autoload :Theme,               "jekyll/theme"
