--- lib/jekyll/site_data.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Jekyll
-  module SiteData
-    autoload :Directory, "jekyll/site_data/directory"
-    autoload :File,      "jekyll/site_data/file"
-  end
-end
